SP5 Blazor: improve component inheritance and naming
