+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.Blazor/BlazorPlot.razor
@@ -1,6 +1,6 @@
 ﻿@using SkiaSharp.Views.Blazor;
 @using SkiaSharp;
-@inherits BlazorPlot
+@inherits BlazorPlotBase
 
 <div>
     <SKCanvasView @ref="SKView" style=@Style OnPaintSurface="OnPaintSurface" IgnorePixelScaling="true"
