+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.Blazor/BlazorPlotGL.razor
@@ -1,6 +1,6 @@
 ﻿@using SkiaSharp.Views.Blazor;
 @using SkiaSharp;
-@inherits BlazorPlot
+@inherits BlazorPlotBase
 
 <div>
     <SKGLView @ref="SKView" style=@Style OnPaintSurface="OnPaintSurface" IgnorePixelScaling="true"
