--- src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.Blazor.WebAssembly/Pages/BlazorPlotDemo.razor
@@ -1,31 +1,0 @@
-﻿@page "/bp"
-@using ScottPlot;
-@using ScottPlot.Blazor;
-
-<p>The <strong>BlazorPlot</strong> manipulates pixel data in memory</p>
-
-<BlazorPlotSkia @ref="BlazorPlot" Style="width: min(100%, 800px); height: 600px;" />
-
-@code {
-    BlazorPlotSkia BlazorPlot = new();
-
-    protected override void OnAfterRender(bool firstRender)
-    {
-        if (firstRender)
-        {
-            BlazorPlot.Plot.Add.Signal(ScottPlot.Generate.Sin());
-            BlazorPlot.Plot.Add.Signal(ScottPlot.Generate.Cos());
-
-            var crosshair = BlazorPlot.Plot.Add.Crosshair(0, 0);
-
-            BlazorPlot.HandlerPointerMoved += (s, e) =>
-            {
-                crosshair.Position = BlazorPlot.GetCoordinates(BlazorPlot.CoordinateToPixel(e));
-                BlazorPlot.Refresh();
-            };
-        }
-
-        base.OnAfterRender(firstRender);
-    }
-
-}
