--- src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.Blazor.WebAssembly/Pages/BlazorPlotGLDemo.razor
@@ -1,31 +1,0 @@
-﻿@page "/bpgl"
-@using ScottPlot;
-@using ScottPlot.Blazor;
-
-<p>The <strong>BlazorPlotGL</strong> uses WebGL for improved performance but may not work in some browsers</p>
-
-<BlazorPlotGL @ref="BlazorPlot" Style="width: min(100%, 800px); height: 600px;" />
-
-@code {
-    BlazorPlotGL BlazorPlot = new();
-
-    protected override void OnAfterRender(bool firstRender)
-    {
-        if (firstRender)
-        {
-            BlazorPlot.Plot.Add.Signal(ScottPlot.Generate.Sin());
-            BlazorPlot.Plot.Add.Signal(ScottPlot.Generate.Cos());
-
-            var crosshair = BlazorPlot.Plot.Add.Crosshair(0, 0);
-
-            BlazorPlot.HandlerPointerMoved += (s, e) =>
-            {
-                crosshair.Position = BlazorPlot.GetCoordinates(BlazorPlot.CoordinateToPixel(e));
-                BlazorPlot.Refresh();
-            };
-        }
-
-        base.OnAfterRender(firstRender);
-    }
-}
-
