--- src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.Blazor.WebAssembly/Pages/GPU.razor
@@ -1,39 +1,0 @@
-﻿@page "/gpu"
-@using ScottPlot;
-@using ScottPlot.Plottables;
-@using ScottPlot.Blazor;
-
-<h1>GPU (WebGL) Canvas</h1>
-
-<p>The canvas below is using WebGL. Better FPS! Don't work on every browser</p>
-
-<div class="container">
-    <div class="row">
-        <div class="col border rounded p-2 canvas-container">
-
-            <BlazorPlotGL @ref="MyChart" Style="width:400px;height:300px;" />
-
-        </div>
-    </div>
-</div>
-
-
-@code {
-    BlazorPlotGL MyChart = new BlazorPlotGL();
-
-    protected override void OnAfterRender(bool firstRender)
-    {
-        if (firstRender)
-        {
-            var crosshair = MyChart.Plot.Add.Crosshair(0, 0);
-            MyChart.HandlerPointerMoved += (s, e) =>
-            {
-
-                crosshair.Position = MyChart.GetCoordinates(MyChart.CoordinateToPixel(e));
-                MyChart.Refresh();
-            };
-        }
-        base.OnAfterRender(firstRender);
-    }
-}
-
