+++ src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.Blazor.WebAssembly/Shared/NavMenu.razor
@@ -1,16 +1,6 @@
 ﻿<ul class="nav flex-column">
     <li class="nav-item px-3">
-        <NavLink class="nav-link fs-5" href="bp">
-            BlazorPlot
-        </NavLink>
-    </li>
-    <li class="nav-item px-3">
-        <NavLink class="nav-link fs-5" href="bpgl">
-            BlazorPlotGL
-        </NavLink>
-    </li>
-    <li class="nav-item px-3">
-        <NavLink class="nav-link fs-5" href="quickstart">
+        <NavLink class="nav-link fs-5" href="/">
             Quickstart
         </NavLink>
     </li>
